+++ libs/@blockprotocol/type-system/crate/src/ontology/url/error.rs
@@ -23,18 +23,14 @@ pub enum ParseBaseUrlError {
 pub enum ParseVersionedUrlError {
     #[error("incorrect formatting")]
     IncorrectFormatting,
-    #[error("missing base url")]
-    MissingBaseUrl,
     #[error("missing version")]
     MissingVersion,
-    #[error("invalid version: {0}")]
-    InvalidVersion(String),
-    #[error("additional end content")]
-    AdditionalEndContent,
+    #[error("invalid version `{0}`: {1}")]
+    InvalidVersion(String, String),
+    #[error("additional end content: {0}")]
+    AdditionalEndContent(String),
     #[error("invalid base url: {0}")]
     InvalidBaseUrl(ParseBaseUrlError),
-    #[error("invalid json: {0}")]
-    InvalidJson(String),
     #[error("URL cannot cannot be more than 2048 characters long")]
     TooLong,
 }